moved alerts steps to seperate page
